rayrender v0.34.0: Add loop subdivision surfaces -Add `subdivision_levels` argument to *_model functions to generate subdivision surfaces -Replace r_obj() mesh with 3D version of R logo generated from SVG file (using raybevel) -`obj_model()` Update MTL to PBR texture conversion -`mesh3d_model()` Remove scale argument -Fix issue where path isn't expanded for meshes -Replace triangle vertex normal with face normal for zero-length normals -Fix ambient light auto option

+++ src/debug.cpp
@@ -429,9 +429,9 @@ void debug_scene(size_t numbercores, size_t nx, size_t ny, size_t ns, int debug_
                                                     max_depth, rng);
                          mat_stack->clear();
                          
-                         routput(i,j) += qr/ns;
-                         goutput(i,j) += qr/ns;
-                         boutput(i,j) += qr/ns;
+                         routput(i,j) += qr/(Float)ns;
+                         goutput(i,j) += qr/(Float)ns;
+                         boutput(i,j) += qr/(Float)ns;
                        }
                      }
                      delete mat_stack;
